+++ civis/ml/tests/test_model.py
@@ -404,6 +404,18 @@ def test_table(mock_res, mock_lt):
                                     index_col=0, client=c)
 
 
+@mock.patch.object(_model, "_load_table_from_outputs")
+@mock.patch.object(_model.ModelFuture, "result")
+@mock.patch.object(_model.ModelFuture, "_set_model_exception", mock.Mock())
+def test_table_None(mock_res, mock_lt):
+    mock_lt.side_effect = FileNotFoundError()
+    c = setup_client_mock(3, 7)
+    mf = _model.ModelFuture(3, 7, client=c)
+    assert mf.table is None
+    mock_lt.assert_called_once_with(3, 7, 'predictions.csv',
+                                    index_col=0, client=c)
+
+
 @mock.patch.object(_model.cio, "file_id_from_run_output",
                    mock.Mock(return_value=11, spec_set=True))
 @mock.patch.object(_model.cio, "file_to_json", return_value='bar')
@@ -528,6 +540,26 @@ def test_metrics_training(mock_file_id_from_run_output):
 
 
 @mock.patch.object(_model.cio, "file_id_from_run_output", autospec=True)
+@mock.patch.object(_model.cio, "file_to_json")
+def test_metrics_training_None(mock_file_to_json,
+                               mock_file_id_from_run_output):
+    mock_file_to_json.return_value = mock.MagicMock(
+        return_value={'metrics': 'foo',
+                      'run': {'status': 'succeeded'}})
+    mock_file_id_from_run_output.return_value = 11
+    c = setup_client_mock(3, 7)
+    mf = _model.ModelFuture(3, 7, client=c)
+    # override validation metadata to be None, as though we ran
+    # a train job without validation
+    mf._val_metadata = None
+
+    mock_file_to_json.return_value = None
+    assert mf.metrics is None
+    mock_file_id_from_run_output.assert_called_with('metrics.json', 3, 7,
+                                                    client=mock.ANY)
+
+
+@mock.patch.object(_model.cio, "file_id_from_run_output", autospec=True)
 @mock.patch.object(_model.cio, "file_to_json",
                    mock.MagicMock(
                        return_value={'metrics': 'foo',
@@ -550,6 +582,31 @@ def mp_setup():
     mock_api = setup_client_mock()
     mp = _model.ModelPipeline('wf', 'dv', client=mock_api)
     return mp
+
+
+@pytest.mark.skipif(not HAS_SKLEARN, reason="scikit-learn not installed")
+def test_modelpipeline_init_err():
+    mock_client = mock.MagicMock()
+    r = Response({'content': None, 'status_code': 9999, 'reason': None})
+    mock_client.templates.get_scripts.side_effect = CivisAPIError(r)
+    with pytest.raises(NotImplementedError):
+        _model.ModelPipeline(LogisticRegression(), 'test',
+                             etl=LogisticRegression(),
+                             client=mock_client)
+    # clean up
+    del _model._NEWEST_CIVISML_VERSION
+
+
+@pytest.mark.skipif(not HAS_SKLEARN, reason="scikit-learn not installed")
+def test_modelpipeline_init_newest():
+    mock_client = mock.MagicMock()
+    mock_client.templates.get_scripts.return_value = {}
+    etl = LogisticRegression()
+    mp = _model.ModelPipeline(LogisticRegression(), 'test', etl=etl,
+                              client=mock_client)
+    assert mp.etl == etl
+    # clean up
+    del _model._NEWEST_CIVISML_VERSION
 
 
 @mock.patch.object(_model, 'ModelFuture')
@@ -718,6 +775,25 @@ def test_modelpipeline_train_from_estimator(mock_ccr, mock_f2c):
 
     est = LogisticRegression()
     mp = _model.ModelPipeline(est, "dv")
+    mock1, mock2 = mock.Mock(), mock.Mock()
+    mock_ccr.return_value = 'res', mock1, mock2
+
+    assert 'res' == mp.train(file_id=7)
+    assert mp.train_result_ == 'res'
+    assert mock_f2c.call_count == 1  # Called once to store input Estimator
+
+
+@pytest.mark.skipif(not HAS_SKLEARN, reason="scikit-learn not installed")
+@mock.patch.object(_model, "APIClient", mock.Mock())
+@mock.patch.object(_model.cio, "file_to_civis")
+@mock.patch.object(_model.ModelPipeline, "_create_custom_run")
+def test_modelpipeline_train_custom_etl(mock_ccr, mock_f2c, mp_setup):
+    # Provide a custom ETL estimator and make sure we can train.
+    mock_api = setup_client_mock()
+    etl = LogisticRegression()
+    mp = _model.ModelPipeline('wf', 'dv', client=mock_api, etl=etl)
+    mock_f2c.return_value = -21
+
     mock1, mock2 = mock.Mock(), mock.Mock()
     mock_ccr.return_value = 'res', mock1, mock2